activity post api수정
