+++ src/apis/boardAPI.js
@@ -3,7 +3,7 @@ import { baseInstance } from "./instance";
 // BulletinBoard 게시판 가져오기
 const getBoardAxios = async (selectBoard) => {
     try {
-      const res = await baseInstance.get(`http://3.39.36.239:8080/boards?name=${selectBoard}`);
+      const res = await baseInstance.get(`/boards?name=${selectBoard}`);
         return res;      
     }catch (error) {
       console.error(error);
@@ -13,7 +13,7 @@ const getBoardAxios = async (selectBoard) => {
 // 게시글 내용 불러오기(수정시)
 const getPostAxios = async (selectPost) => {
   try{
-    const res = await baseInstance.get(`http://3.39.36.239:8080/posts?id=${selectPost}`);
+    const res = await baseInstance.get(`/posts?id=${selectPost}`);
       return res.data.payload;
   }catch(error){
     console.log(error);
@@ -23,7 +23,7 @@ const getPostAxios = async (selectPost) => {
 // 게시글 등록하기
 const setWritePostAxios = async (data) => {
   try{
-    const res = await baseInstance.post("http://3.39.36.239:8080/posts", data);
+    const res = await baseInstance.post("/posts", data);
       return res;
     }catch(error){
       console.log(error);
@@ -33,7 +33,7 @@ const setWritePostAxios = async (data) => {
 // 게시글 수정하기
 const setEditedPostAxios = async (data) => {
   try{
-    const res = await baseInstance.patch("http://3.39.36.239:8080/posts", data);
+    const res = await baseInstance.patch("/posts", data);
       return res;
     }catch(error){
       console.log(error);
@@ -43,7 +43,7 @@ const setEditedPostAxios = async (data) => {
 // 게시글 삭제하기
 const deletePostAxios = async (postId) => {
   try{
-    const res = await baseInstance.delete(`http://3.39.36.239:8080/posts/${postId}`)
+    const res = await baseInstance.delete(`/posts/${postId}`)
       console.log(res)
       return res;
   }catch(error){
