+++ src/components/ImageSlider.jsx
@@ -1,6 +1,6 @@
 import Slider from "react-slick";
-import home from "../../assets/images/home.png";
-import home2 from "../../assets/images/home2.png";
+import home from "../assets/images/home.png";
+import home2 from "../assets/images/home2.png";
 import styled from "styled-components";
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
