+++ src/components/activity/ActivityUserModal.jsx
@@ -1,4 +1,4 @@
-import {useState} from "react";
+import { useState } from "react";
 import styled from "styled-components";
 
 const ActivityUserModal  = (props) =>{
